Add check for empty string in database define
We had an ambigious error that was difficult to track down the cause:
```
Error: Could not retrieve catalog from remote server: Error 400 on SERVER: Failed to parse inline template: undefined method `gsub' for nil:NilClass at /usr/share/puppet/production/current/modules/govuk/manifests/app/envvar/database_url.pp:32 on node i-xxxxxxxxxxxxx
Warning: Not using cache on failed catalog
Error: Could not retrieve catalog; skipping run
```

This was the extent of the error, and did not show what was using the
defined type that caused the error. We eventually tracked this down to
an empty missing value in hieradata, but an empty string seemed to be
passed in which meant that Puppet itself did not error with a missing
parameter in the defined type, but the function threw an error due to an
empty string.

We're still not 100% clear on the cause, but we're hoping that a check
for an empty string will make this issue clearer in the future and avoid
the function throwing the error rather than Puppet code.