Add variable for integration database password
We need to pass the integration database password through to the
Whitehall database sanitise and sync script.

Set the password and pass it as a second argument to
`sanitise-and-sync-db.sh`.

At the same time, disable printing the commands to the Jenkins console
so we don't leak sensitive information on screen.