Remove logspout in favour of AWS implementation
Logspout was being incorrectly installed on some machines and not
installed on others. This is because CI/management machines match the
conditional, while actual integration machines have AWS domains and do
not match. The comment suggests this code is out of date, as logit is
actually available in staging and production.

For now, it seems best to remove this feature until we finish moving to
AWS and can make it work in all environments.